requirements.txt
streamlit==1.32.0
langchain==0.0.353
langchain_community==0.0.20
pillow==10.1.0
streamlit_supabase_auth
langchain-ollama
langchain_core